requirements.txt
asttokens==2.0.4
atomicwrites==1.4.0
attrs==20.3.0
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
colorama==0.4.4
configparser==5.0.2
executing==0.5.4
flake8==3.9.1
freezegun==1.1.0
greenlet==1.0.0
icecream==2.1.0
iniconfig==1.1.1
mccabe==0.6.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.10.1
pycryptodomex==3.10.1
pyflakes==2.3.1
Pygments==2.8.1
pyparsing==2.4.7
PyQt5==5.15.2
pyqt5-plugins==5.15.2.2.1.0
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
PyQt5-stubs==5.15.2.0
pyqt5-tools==5.15.2.3.0.2
pytest==6.2.2
python-dateutil==2.8.1
python-dotenv==0.17.0
PyYAML==5.4.1
qt5-applications==5.15.2.2.1
qt5-tools==5.15.2.1.0.1
six==1.15.0
SQLAlchemy==1.4.5
structlog==21.1.0
tabulate==0.8.9
toml==0.10.2
